requirements.txt
numpy
Biopython
pandas
matplotlib

